♻️ refactor: text br 태그 추가 및 리펙토링 작업
