+++ package.json
@@ -19,6 +19,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
+    "react-spinners": "^0.15.0",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
