--- src/constants/text.ts
@@ -1,3 +1,0 @@
-// constants/text.ts
-
-export const INITIAL_TEXT = '샘플 텍스트입니다. SNAPSUM은 AI를 활용한 숏폼 영상 제작 서비스로 누구나 쉽게 영상 크리에이터로의 전환을 돕습니다. 먼저 링크를 불여넣고, 업로드하고자 하는 플랫폼을 선택 선택합니다. 링크에서 그를 원하는 텍스트를 AI가 분석하고, 요약한 텍스트를 사용자가 자유롭게 문단으로 분리할 수 있습니다. 분리한 문단 기준으로 이미지를 생성하고, 선택한 TTS를 합쳐 영상을 제작하게 됩니다.';
